+++ main.go
@@ -17,7 +17,7 @@ import (
 )
 
 const (
-	version = "1.0.4" // release-arvigo-backend-1.0.4
+	version = "1.0.5" // release-arvigo-backend-1.0.5
 	appName = "arvigo-backend"
 )